+++ R/WorldClim_data.R
@@ -17,53 +17,23 @@
 #' @param variable Allows to specify which variables you want to retrieve Possible entries are:
 #' "tmax","tmin","prec" and/or "bioc".
 #' @param year Specify the year you want to retrieve data. Possible entries are:
-#' "2030", "2050", "2070" and/or "2090". You can  use a vector to provide more than one entry.
+#' "2030", "2050", "2070" and/or "2090". You can use a vector to provide more than one entry.
 #' @param gcm GCMs to be considered in future scenarios. You can use a vector to provide more than
 #' one entry.
-#'  | **CODE** | **GCM** |
-#'  | ---- | ---------------- |
-#'  | ac  | ACCESS-CM2 |
-#'  | ae  | ACCESS-ESM1-5 |
-#'  | bc  | BCC-CSM2-MR |
-#'  | ca  | CanESM5 |
-#'  | cc  | CanESM5-CanOE |
-#'  | ce  | CMCC-ESM2 |
-#'  | cn  | CNRM-CM6-1 |
-#'  | ch  | CNRM-CM6-1-HR |
-#'  | cr  | CNRM-ESM2-1 |
-#'  | ec  | EC-Earth3-Veg |
-#'  | ev  | EC-Earth3-Veg-LR |
-#'  | fi  | FIO-ESM-2-0 |
-#'  | gf  | GFDL-ESM4 |
-#'  | gg  | GISS-E2-1-G |
-#'  | gh  | GISS-E2-1-H |
-#'  | hg  | HadGEM3-GC31-LL |
-#'  | in  | INM-CM4-8 |
-#'  | ic  | INM-CM5-0 |
-#'  | ip  | IPSL-CM6A-LR |
-#'  | me  | MIROC-ES2L |
-#'  | mi  | MIROC6 |
-#'  | mp  | MPI-ESM1-2-HR |
-#'  | ml  | MPI-ESM1-2-LR |
-#'  | mr  | MRI-ESM2-0 |
-#'  | uk  | UKESM1-0-LL |
-#' @md
 #' @param ssp SSPs for future data. Possible entries are: "126", "245", "370" and/or "585".
 #' You can use a vector to provide more than one entry.
 #' @param resolution You can select one resolution from the following alternatives: 10, 5, 2.5 OR 30.
 #'
-#' @details This function will create a folder entitled "input_data/WorldClim_data_current" or
-#' "input_data/WorldClim_data_future". All the data downloaded will be stored in this folder.
-#' Note that, despite being possible to retrieve a lot of data at once, it is not recommended to do
-#' so, since the data is very heavy.
+#' @details This function will create a folder. All the data downloaded will be stored in this
+#' folder. Note that, despite being possible to retrieve a lot of data at once, it is not
+#' recommended to do so, since the data is very heavy.
 #'
-#' @returns If data is not downloaded, the function downloads the data and has no return value. If
-#' the data is downloaded, it imports the data as a \code{stack}.
+#' @returns If data is not downloaded, the function downloads the data and has no return value.
 #'
-#' @references https://www.worldclim.org/data/index.html
+#' @references [https://www.worldclim.org/data/index.html](https://www.worldclim.org/data/index.html)
 #'
 #' @author Luíz Fernando Esser (luizesser@gmail.com)
-#' https://luizfesser.wordpress.com
+#' [https://luizfesser.wordpress.com](https://luizfesser.wordpress.com)
 #'
 #' @examples
 #' \donttest{
@@ -85,30 +55,35 @@
 #'                resolution = 10)
 #' }
 #'
-#' @importFrom httr GET write_disk http_error
-#' @importFrom raster stack
-#' @importFrom gtools mixedsort
-#' @importFrom cli cli_alert_warning cli_abort
+#' @importFrom httr2 request req_url req_error req_perform resp_is_error resp_status
+#' @importFrom cli cli_alert_warning cli_abort cli_inform
 #' @importFrom utils unzip
 #'
 #' @export
+WorldClim_data <- function(path = NULL,
+                           period = "current",
+                           variable = "bioc",
+                           year = "2090",
+                           gcm = "mi",
+                           ssp = "585",
+                           resolution = 10) {
 
-WorldClim_data <- function(path = NULL, period = "current", variable = "bioc", year = "2090", gcm = "mi", ssp = "585", resolution = 10) {
+  ## argument checks (unchanged)
   if (!all(period %in% c("current", "future"))) {
     cli::cli_abort(c(
-      "x" = "Assertion on {.var period} failed. ",
+      "x" = "Assertion on {.var period} failed.",
       "i" = "{.var period} must be element of set ['current', 'future']."
     ))
   }
   if (!all(variable %in% c("bioc", "tmax", "tmin", "prec"))) {
     cli::cli_abort(c(
-      "x" = "Assertion on {.var variable} failed. ",
+      "x" = "Assertion on {.var variable} failed.",
       "i" = "{.var variable} must be element of set ['bioc', 'tmax','tmin','prec']."
     ))
   }
   if (!all(year %in% c("2030", "2050", "2070", "2090"))) {
     cli::cli_abort(c(
-      "x" = "Assertion on {.var year} failed. ",
+      "x" = "Assertion on {.var year} failed.",
       "i" = "{.var year} must be element of set ['2030', '2050', '2070', '2090']."
     ))
   }
@@ -118,20 +93,20 @@ WorldClim_data <- function(path = NULL, period = "current", variable = "bioc", y
     "mr", "uk", "all"
   ))) {
     cli::cli_abort(c(
-      "x" = "Assertion on {.var gcm} failed. ",
+      "x" = "Assertion on {.var gcm} failed.",
       "i" = "{.var gcm} must be element of set ['ac','ae','bc','ca','cc','ce','cn','ch','cr','ec',
-      'ev','fi','gf','gg','gh','hg','in','ic','ip','me','mi','mp','ml','mr','uk', 'all']."
+      'ev','fi','gf','gg','gh','hg','in','ic','ip','me','mi','mp','ml','mr','uk','all']."
     ))
   }
   if (!all(ssp %in% c("126", "245", "370", "585"))) {
     cli::cli_abort(c(
-      "x" = "Assertion on {.var ssp} failed. ",
+      "x" = "Assertion on {.var ssp} failed.",
       "i" = "{.var ssp} must be element of set ['126', '245', '370', '585']."
     ))
   }
   if (!all(resolution %in% c(10, 5, 2.5, 30))) {
     cli::cli_abort(c(
-      "x" = "Assertion on {.var resolution} failed. ",
+      "x" = "Assertion on {.var resolution} failed.",
       "i" = "{.var resolution} must be element of set [10, 5, 2.5, 30]."
     ))
   }
@@ -147,6 +122,36 @@ WorldClim_data <- function(path = NULL, period = "current", variable = "bioc", y
   res <- ifelse(resolution == 30, "s", "m")
   l <- list()
 
+  .warn <- function(...) cli::cli_alert_warning(paste0(...))
+
+  ## helper: GET file with httr2, but never throw on HTTP status
+  .download_file_httr2 <- function(url, destfile) {
+    req <- httr2::request(url) |>
+      httr2::req_error(is_error = \(resp) FALSE)  # keep 4xx/5xx as responses, not errors [web:33][web:38]
+
+    resp <- try(httr2::req_perform(req), silent = TRUE)
+
+    if (inherits(resp, "try-error")) {
+      .warn("Failed to perform request to ", url, ".")
+      return(FALSE)
+    }
+
+    if (httr2::resp_is_error(resp)) {
+      .warn("HTTP error ", httr2::resp_status(resp), " when requesting ", url, ".")
+      return(FALSE)
+    }
+
+    # write raw body to disk; works for .zip and .tif [web:27][web:28]
+    body <- httr2::resp_body_raw(resp)
+    ok <- try(writeBin(body, destfile), silent = TRUE)
+    if (inherits(ok, "try-error")) {
+      .warn("Failed to write response body to file ", destfile, ".")
+      return(FALSE)
+    }
+    TRUE
+  }
+
+  ## CURRENT PERIOD
   if (period == "current") {
     if (is.null(path)) {
       path <- "input_data/WorldClim_data_current"
@@ -154,29 +159,37 @@ WorldClim_data <- function(path = NULL, period = "current", variable = "bioc", y
     if (!dir.exists(path)) {
       dir.create(path, recursive = TRUE)
     }
-    if (length(list.files(path, pattern = ".tif$", full.names = T)) == 0) {
-      print(paste0("current_", resolution, res))
-      httr::GET(
-        url = paste0(
-          "https://geodata.ucdavis.edu/climate/worldclim/2_1/base/wc2.1_",
-          resolution,
-          res, "_bio.zip"
-        ),
-        httr::write_disk(paste0(path, "/current_", resolution, res, ".zip"))
+
+    tif_files <- list.files(path, pattern = ".tif$", full.names = TRUE)
+
+    if (length(tif_files) == 0) {
+      url <- paste0(
+        "https://geodata.ucdavis.edu/climate/worldclim/2_1/base/wc2.1_",
+        resolution, res, "_bio.zip"
       )
-      utils::unzip(
-        zipfile = paste0(path, "/current_", resolution, res, ".zip"),
-        exdir = path
-      )
+      zipfile <- file.path(path, paste0("current_", resolution, res, ".zip"))
+
+      cli::cli_inform(c("Downloading WorldClim current data from {.url {url}}"))
+
+      ok <- .download_file_httr2(url, zipfile)
+      if (!ok) {
+        .warn("Failed to download WorldClim current data. The server may be unavailable or the URL may have changed.")
+        return(invisible(l))
+      }
+
+      uz <- try(utils::unzip(zipfile, exdir = path), silent = TRUE)
+      if (inherits(uz, "try-error")) {
+        .warn("Download succeeded but unzip failed for file ", zipfile, ".")
+        return(invisible(l))
+      }
+
+      tif_files <- list.files(path, pattern = ".tif$", full.names = TRUE)
     } else {
-      print(paste0("The file for current scenario is already downloaded."))
-      print(paste0("Importing current from folder..."))
-      l[["current"]] <- raster::stack(list.files(path, pattern = ".tif$", full.names = T))
-      l[["current"]] <- l[["current"]][[gtools::mixedsort(names(l[["current"]]))]]
-      names(l[["current"]]) <- paste0("bio", 1:19)
+      cli::cli_inform("The file for current scenario is already downloaded.")
     }
   }
 
+  ## FUTURE PERIOD
   if (period == "future") {
     if (is.null(path)) {
       path <- "input_data/WorldClim_data_future"
@@ -184,6 +197,7 @@ WorldClim_data <- function(path = NULL, period = "current", variable = "bioc", y
     if (!dir.exists(path)) {
       dir.create(path, recursive = TRUE)
     }
+
     all_gcm <- c(
       "ac", "ae", "bc", "ca", "cc", "ce", "cn", "ch", "cr", "ec", "ev", "fi",
       "gf", "gg", "gh", "hg", "in", "ic", "ip", "me", "mi", "mp", "ml",
@@ -201,48 +215,39 @@ WorldClim_data <- function(path = NULL, period = "current", variable = "bioc", y
       gcm <- all_gcm
     }
     gcm3 <- gcm2[match(gcm, all_gcm)]
+
     all_year <- c("2030", "2050", "2070", "2090")
     year2 <- c("2021-2040", "2041-2060", "2061-2080", "2081-2100")
     year3 <- year2[match(year, all_year)]
-    for (g in 1:length(gcm)) {
-      for (s in 1:length(ssp)) {
-        for (y in 1:length(year)) {
-          if (!file.exists(paste0(path, "/", gcm[g], "_ssp", ssp[s], "_", resolution, "_", year[y], ".tif"))) {
-            print(paste0(gcm[g], "_ssp", ssp[s], "_", resolution, "_", year[y]))
-            if (!httr::http_error(paste0(
-              "https://geodata.ucdavis.edu/cmip6/", resolution,
-              res, "/", gcm3[g], "/ssp", ssp[s], "/wc2.1_", resolution,
-              res, "_", variable, "_", gcm3[g], "_ssp", ssp[s], "_",
-              year3[y], ".tif"
-            ))) {
-              try(httr::GET(
-                url = paste0(
-                  "https://geodata.ucdavis.edu/cmip6/", resolution,
-                  res, "/", gcm3[g], "/ssp", ssp[s], "/wc2.1_", resolution,
-                  res, "_", variable, "_", gcm3[g], "_ssp", ssp[s], "_",
-                  year3[y], ".tif"
-                ),
-                httr::write_disk(paste0(
-                  path, "/", gcm[g], "_ssp", ssp[s],
-                  "_", resolution, "_", year[y], ".tif"
-                ))
-              ))
+
+    for (g in seq_along(gcm)) {
+      for (s in seq_along(ssp)) {
+        for (y in seq_along(year)) {
+          nome <- paste0(gcm[g], "_ssp", ssp[s], "_", resolution, "_", year[y])
+          destfile <- file.path(path, paste0(nome, ".tif"))
+
+          if (!file.exists(destfile)) {
+            url <- paste0(
+              "https://geodata.ucdavis.edu/cmip6/", resolution, res, "/",
+              gcm3[g], "/ssp", ssp[s], "/wc2.1_", resolution, res, "_",
+              variable, "_", gcm3[g], "_ssp", ssp[s], "_", year3[y], ".tif"
+            )
+
+            cli::cli_inform(c("Downloading WorldClim future data: {nome} from {.url {url}}"))
+
+            ok <- .download_file_httr2(url, destfile)
+            if (!ok) {
+              .warn("Failed to download ", nome, ". The server may be unavailable or the URL may have changed.")
+              next
             }
           } else {
-            print(paste0(
-              "The file for future scenario (",
-              paste0(path, "/", gcm[g], "_ssp", ssp[s], "_", resolution, "_", year[y], ".tif"),
-              ") is already downloaded."
+            cli::cli_inform(paste0(
+              "The file for future scenario (", destfile,") is already downloaded."
             ))
-            nome <- paste0(gcm[g], "_ssp", ssp[s], "_", resolution, "_", year[y])
-            print(paste0("Importing ", nome, " from folder..."))
-            l[[nome]] <- raster::stack(list.files(path, pattern = paste0(nome, ".tif$"), full.names = T))
-            l[[nome]] <- l[[nome]][[gtools::mixedsort(names(l[[nome]]))]]
-            names(l[[nome]]) <- paste0("bio", 1:19)
           }
         }
       }
     }
   }
-  return(l)
+  return(invisible(l))
 }